Factored AdapterIt
